+++ packages/components/src/presentation-components/properties/inputs/QuantityPropertyEditorInput.tsx
@@ -44,7 +44,7 @@ QuantityPropertyEditorInput.displayName = "QuantityPropertyEditorInput";
 type QuantityPropertyValueInputProps = QuantityPropertyEditorImplProps & UseQuantityValueInputProps;
 
 const QuantityPropertyValueInput = forwardRef<PropertyEditorAttributes, QuantityPropertyValueInputProps>(
-  ({ propertyRecord, onCommit, koqName, schemaContext, initialRawValue, setFocus }, ref) => {
+  ({ propertyRecord, onCommit, koqName, schemaContext, initialRawValue, setFocus, onCancel }, ref) => {
     const { quantityValue, inputProps } = useQuantityValueInput({ koqName, schemaContext, initialRawValue });
     const [isEditing, setEditing] = useState(false);
     const value = isEditing ? quantityValue.highPrecisionFormattedValue : quantityValue.defaultFormattedValue;
@@ -83,6 +83,16 @@ const QuantityPropertyValueInput = forwardRef<PropertyEditorAttributes, Quantity
       }
     }, [inputProps.disabled, setFocus]);
 
+    const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+      if (e.key === "Escape") {
+        onCancel?.();
+      }
+      if (e.key === "Enter") {
+        inputRef.current?.blur();
+        e.stopPropagation();
+      }
+    };
+
     return (
       <Input
         {...inputProps}
@@ -96,8 +106,11 @@ const QuantityPropertyValueInput = forwardRef<PropertyEditorAttributes, Quantity
         }}
         onFocus={() => {
           setEditing(true);
-          inputRef.current?.setSelectionRange(0, 9999);
+          requestAnimationFrame(() => {
+            inputRef.current && inputRef.current.setSelectionRange(0, quantityValue.highPrecisionFormattedValue === inputProps.placeholder ? 0 : 9999);
+          });
         }}
+        onKeyDown={handleKeyDown}
       />
     );
   },