+++ packages/components/src/test/properties/inputs/QuantityPropertyEditorInput.test.tsx
@@ -11,6 +11,7 @@ import { BeUiEvent } from "@itwin/core-bentley";
 import { IModelApp } from "@itwin/core-frontend";
 import { Format, FormatType, ParseError } from "@itwin/core-quantity";
 import { KoqPropertyValueFormatter } from "@itwin/presentation-common";
+import { stubRaf } from "../../_helpers/Common.js";
 import { createTestPropertyRecord } from "../../_helpers/UiComponents.js";
 import { SchemaMetadataContextProvider } from "../../../presentation-components/common/SchemaMetadataContext.js";
 import { QuantityEditorName } from "../../../presentation-components/properties/editors/QuantityPropertyEditor.js";
@@ -35,6 +36,7 @@ describe("<QuantityPropertyEditorInput />", () => {
   const format = new Format("test format");
   const formatterSpec = {
     applyFormatting: sinon.stub<[number], string>(),
+    unitConversions: [{ name: "test unit", label: "unit" }],
     format,
   };
   const parserSpec = {
@@ -47,6 +49,8 @@ describe("<QuantityPropertyEditorInput />", () => {
     ReturnType<KoqPropertyValueFormatter["getFormatterSpec"]>
   >;
   let getParserSpecStub: sinon.SinonStub<Parameters<KoqPropertyValueFormatter["getParserSpec"]>, ReturnType<KoqPropertyValueFormatter["getParserSpec"]>>;
+
+  stubRaf();
 
   before(() => {
     getFormatterSpecStub = sinon.stub(KoqPropertyValueFormatter.prototype, "getFormatterSpec");
@@ -139,7 +143,7 @@ describe("<QuantityPropertyEditorInput />", () => {
 
   it("allows entering quantity value when schema context is available", async () => {
     const ref = createRef<PropertyEditorAttributes>();
-    const spy = sinon.stub<Parameters<Required<PropertyEditorProps>["onCommit"]>, ReturnType<Required<PropertyEditorProps>["onCommit"]>>();
+    const spy = sinon.spy();
     const record = createRecord({ initialValue: undefined, kindOfQuantityName: "TestKOQ" });
     const { getByRole, user } = render(
       <SchemaMetadataContextProvider imodel={{} as IModelConnection} schemaContextProvider={() => schemaContext}>
@@ -147,12 +151,15 @@ describe("<QuantityPropertyEditorInput />", () => {
       </SchemaMetadataContextProvider>,
     );
 
-    const input = await waitFor(() => getByRole("textbox"));
-    await waitFor(() => expect((input as HTMLInputElement).disabled).to.be.false);
+    const input = await waitFor(() => getByRole("textbox") as HTMLInputElement);
+    await waitFor(() => expect(input.disabled).to.be.false);
 
-    await user.type(input, "123.4 unit");
+    // Verify that selection logic is applied
+    await waitFor(() => expect(input.selectionEnd).to.eq(0));
+    await user.type(input, "123.4 ", { skipClick: true });
     await user.tab();
 
+    await waitFor(() => expect(input.value).to.eq("123.4 unit"));
     await waitFor(() => {
       expect(spy).to.be.calledWith({
         propertyRecord: record,
@@ -174,7 +181,7 @@ describe("<QuantityPropertyEditorInput />", () => {
   });
 
   it("should focus on input if setFocus is true", async () => {
-    const record = createRecord({ initialValue: undefined, kindOfQuantityName: "TestKOQ" });
+    const record = createRecord({ initialValue: 10, kindOfQuantityName: "TestKOQ" });
     const ref = createRef<PropertyEditorAttributes>();
 
     const { getByRole } = render(
@@ -183,9 +190,42 @@ describe("<QuantityPropertyEditorInput />", () => {
       </SchemaMetadataContextProvider>,
     );
 
-    const input = await waitFor(() => getByRole("textbox"));
+    const input = (await waitFor(() => getByRole("textbox"))) as HTMLInputElement;
     await waitFor(() => {
       expect(input).to.be.eq(document.activeElement);
     });
+  });
+
+  it("commits value and blurs input when Enter is pressed", async () => {
+    const record = createRecord({ initialValue: 10, kindOfQuantityName: "TestKOQ" });
+    const onCommitSpy = sinon.spy();
+    const { getByRole, user } = render(
+      <SchemaMetadataContextProvider imodel={{} as IModelConnection} schemaContextProvider={() => schemaContext}>
+        <QuantityPropertyEditorInput propertyRecord={record} onCommit={onCommitSpy} />
+      </SchemaMetadataContextProvider>,
+    );
+    const inputContainer = (await waitFor(() => getByRole("textbox"))) as HTMLInputElement;
+
+    await user.click(inputContainer);
+    await waitFor(() => expect(inputContainer.selectionEnd).to.eq(inputContainer.value.length));
+    await user.keyboard("{Enter}");
+
+    expect(onCommitSpy).to.be.calledOnce;
+  });
+
+  it("calls onCancel when Escape is pressed", async () => {
+    const record = createRecord({ initialValue: 10, kindOfQuantityName: "TestKOQ" });
+    const onCancelSpy = sinon.spy();
+    const { getByRole, user } = render(
+      <SchemaMetadataContextProvider imodel={{} as IModelConnection} schemaContextProvider={() => schemaContext}>
+        <QuantityPropertyEditorInput propertyRecord={record} onCancel={onCancelSpy} />
+      </SchemaMetadataContextProvider>,
+    );
+    const inputContainer = await waitFor(() => getByRole("textbox"));
+
+    await user.click(inputContainer);
+    await user.keyboard("{Escape}");
+
+    expect(onCancelSpy).to.be.calledOnce;
   });
 });