+++ packages/hierarchies-react/CHANGELOG.md
@@ -786,6 +786,16 @@
 
 - [#847](https://github.com/iTwin/presentation/pull/847): Moving tree rendering components to a new design systems.
 
+## 1.9.12
+
+### Patch Changes
+
+- [#1215](https://github.com/iTwin/presentation/pull/1215): Update dependencies.
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.7.10
+  - @itwin/presentation-shared@1.2.9
+  - @itwin/unified-selection@1.6.7
+
 ## 1.9.11
 
 ### Patch Changes