+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log - @itwin/presentation-testing
 
+## 5.4.15
+
+### Patch Changes
+
+- [#1215](https://github.com/iTwin/presentation/pull/1215): Update dependencies.
+- Updated dependencies:
+  - @itwin/presentation-components@5.13.0
+
 ## 5.4.14
 
 ### Patch Changes